Consolidate rpc server/client files in morpheusvm
